requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.6
django-allauth==0.52.0
django-crispy-forms==1.14.0
django_debug_toolbar==3.8.1
h11==0.14.0
idna==3.4
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
platformdirs==2.6.2
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
yarl==1.8.2
